Week6Activity8/requirements.txt
# Currency Exchange CLI - Dependencies
# =====================================

# HTTP requests library for API calls
requests>=2.28.0

# Environment variable management
python-dotenv>=1.0.0

# Rich CLI library for beautiful terminal output
rich>=13.0.0

# Built-in modules (no installation needed):
# - sqlite3: Database operations
# - logging: Logging system
# - datetime: Timestamp handling
